Make Dockerfile runnable.
